+++ ng_trajectory/segmentators/utils.py
@@ -100,6 +100,17 @@ def mapCreate(
     return MAP, MAP_ORIGIN, MAP_GRID
 
 
+def getMap() -> numpy.ndarray:
+    """Obtain the current map representation.
+
+    Returns:
+    MAP -- current map, numpy.ndarray
+    """
+    global MAP
+
+    return MAP
+
+
 def pointInBounds(point: list) -> bool:
     """Check whether a point is inside the map bounds.
 
@@ -128,6 +139,9 @@ def pointToMap(point: list) -> numpy.ndarray:
     """
     global MAP_ORIGIN, MAP_GRID
 
+    # If the point is outside, might return a value like
+    # numpy.round(...) = -1, which yields a totally insane
+    # number when converted to uint64.
     return numpy.round(
         numpy.subtract(numpy.asarray(point)[:2], MAP_ORIGIN) / MAP_GRID
     ).astype(numpy.uint64)