+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
     - _Profile_
         - Parameter 'friction_map' to load a friction map from '.npy' file.
         - Use friction for computing maximum velocity from the friction map if available.
+        - Parameter 'friction_map_inverse' to invert the values in the friction map.
 - Segmentators
     - Function `getMap` to obtain current internal map used by the segmentators.
 - 'ng_generate_data'