+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
         - Parameter 'friction_map' to load a friction map from '.npy' file.
         - Use friction for computing maximum velocity from the friction map if available.
         - Parameter 'friction_map_inverse' to invert the values in the friction map.
+        - Support for zero friction (impassable area).
 - Segmentators
     - Function `getMap` to obtain current internal map used by the segmentators.
 - 'ng_generate_data'