+++ CHANGELOG.md
@@ -4,6 +4,35 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Unreleased
+## 1.11.0 - 2023-11-21
+### Added
+- Criterions
+    - _Profile_
+        - Parameter 'friction_map' to load a friction map from '.npy' file.
+        - Use friction for computing maximum velocity from the friction map if available.
+        - Parameter 'friction_map_inverse' to invert the values in the friction map.
+        - Support for zero friction (impassable area).
+        - CSV file contains progress along the centerline [m].
+- Segmentators
+    - Function `getMap` to obtain current internal map used by the segmentators.
+- 'ng_generate_data'
+    - Parameter `--friction-map` to create a friction map from an image.
+    - Parameter `--fonly/--friction-map-only` to skip generating map data and generate only friction map data.
+    - Parameter `--friction-map-inverse` to inverse the colors in the friction map image.
+
+### Changed
+- Criterions
+    - _Profile_
+        - Time difference between points is constrained to 10s.
+        - Generated CSV file holds lap time and path length with the first trajectory point.
+
+### Fixed
+- Criterions
+    - _Profile_
+        - Avoid crashing when logfile is not given.
+- Segmentators
+    - Catch and handle OverflowError when identifying invalid points.
+
 ## 1.10.0 - 2023-11-15
 ### Changed
 - Update all files to mitigate majority of the flake8 errors.